+++ packages/subgraph/tests/dao/dao.test.ts
@@ -331,7 +331,9 @@ test('Run dao (handleExecuted) for Token mappings with mock event', () => {
   createDaoEntityState(daoEntity, ADDRESS_ONE, DAO_TOKEN_ADDRESS);
 
   let proposalId =
-    Address.fromHexString(CONTRACT_ADDRESS).toHexString() + '_' + '0x0';
+    Address.fromHexString(CONTRACT_ADDRESS).toHexString() +
+    '_' +
+    '0x0000000000000000000000000000000000000000000000000000000000000000';
 
   createTokenVotingProposalEntityState();
 
@@ -345,7 +347,7 @@ test('Run dao (handleExecuted) for Token mappings with mock event', () => {
   let actions = createDummyActions(DAO_ADDRESS, '0', callData);
   let event = createNewExecutedEvent(
     Address.fromHexString(CONTRACT_ADDRESS).toHexString(),
-    '0',
+    '0x0000000000000000000000000000000000000000000000000000000000000000',
     actions,
     [Bytes.fromUTF8('')],
     Address.fromHexString(DAO_ADDRESS).toHexString()